Renamed EditRetryDialog2 back to EditRetryDialog
This is part of the cleanup after introducing v2
